+++ projects/js-packages/script-data/CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2025-02-24
+### Added
+- Update the site host utility functions. [#41713]
+- User data: Add permissions to the current user object. [#41859]
+
+### Changed
+- Script data: Remove unneeded specific host check. [#41890]
+
 ## [0.2.0] - 2025-02-11
 ### Added
-- Added some helper funcitons for site types. [#41393]
+- Add some helper functions for site types. [#41393]
 
 ## [0.1.9] - 2025-01-23
 ### Changed
@@ -49,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added jetpack-script-data package to consolidate the logic for Jetpack Initial state [#38430]
 
+[0.2.1]: https://github.com/Automattic/jetpack-script-data/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/Automattic/jetpack-script-data/compare/v0.1.9...v0.2.0
 [0.1.9]: https://github.com/Automattic/jetpack-script-data/compare/v0.1.8...v0.1.9
 [0.1.8]: https://github.com/Automattic/jetpack-script-data/compare/v0.1.7...v0.1.8