+++ projects/packages/ip/CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2025-02-24
+### Changed
+- Update dependencies.
+
 ## [0.4.1] - 2024-11-25
 ### Changed
-- Updated package dependencies. [#40258]
+- Update package dependencies. [#40258]
 
 ## [0.4.0] - 2024-11-14
 ### Removed
@@ -70,6 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add jetpack-ip package functionality [#28846]
 - Initialized the package. [#28765]
 
+[0.4.2]: https://github.com/automattic/jetpack-ip/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/automattic/jetpack-ip/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/automattic/jetpack-ip/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/automattic/jetpack-ip/compare/v0.3.0...v0.3.1